Clarify the first scenario's use case
Lenny and I were both confused on this since it originally said to select an artist, but the 'new song' button is already on an artist's page and per the previous iteration, the new song will automatically be created for this artist. Therefore, since we are trying to test that you cannot create a new song without a song title, this makes it clearer.